Estimation Issues
This code includes an untested way to plot extra values on a g2(0) chart.
However, calc.py is in the midst of refurbishment.
If g2(0) = (amp_mpe - bg)/(amp_other - bg), the mean of g2(0) is not this function with variables replaced by their means.
Essentially, a 2D distribution of g2(0) must be calculated.