Restructure
Newly available raw data has triggered a restructure of how it is stored in the 'data' folder.
Saves and results will now be produced within identical subfolder hierarchies.
Additionally, ml_fedot.py is an attempt to use the AutoML FEDOT package for prediction, although with no inspirational success.
The likelihood.py file is an abandoned attempt to mimic a maximum a posteriori probability (MAP) estimator, ultimately deemed unnecessary for now as it is just another form of g2(0) estimation.
This commit is prior to any broad dataset cleaning or small-sample least-squares fitting.