Standard Error Issues
It is possible to make parameter estimates with both a least-squares residual, i.e. (fit-data)^2, and a Poisson-based likelihood method residual from the Cortes paper, i.e. fit-data*log(fit).
The lmfit package appears to apply a sum of squares to the objective function before fitting, so both are square-rooted in the actual objective function.
However, the standard error returned from the latter residual seems consistently worse than from the former, contradicting assertions from the Cortes paper.
This needs to be investigated.