Doubt Moment
The function fitted to two-photon events has thus far been considered a probability distribution, with two subsequent choices:
- Data is normalised prior to fitting.
- Inverse sampling transform is used to 'uniformly' generate data.
Unfortunately, a closer read of the Cortes paper suggests that the function actually denotes the expected number of photon events to see with any delay after a point in time, i.e. the expectation value of a Poisson distribution.
This means that both choices are above are in doubt, with normalisation needing to be reconsidered and inverse sampling outright wrong.