Reviewed Paper Update
Following the review of the paper, now accepted and titled "Harnessing data augmentation to quantify uncertainty in the early estimation of single-photon source quality" (DOI: 10.1088/2632-2153/ad0d11), this commit updates the README, as well as mc.py and plot_paper.py.
Basically, expert users can now overwrite g and bg in the mc.py script, allowing replication of the 'low g' figure added to the manuscript.